Move startup to the base class
